requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
